% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downlit-html.R
\name{downlit_html}
\alias{downlit_html}
\alias{downlit_xml_node}
\title{Automatically link references and articles in an HTML page}
\usage{
downlit_html(input_path, output_path)

downlit_xml_node(node)
}
\arguments{
\item{input_path, output_path}{Input and output paths for HTML file}

\item{node}{An \code{xml2::xml_node}}
}
\value{
Invisibly returns \code{output_path}.
}
\description{
The autolinker has two components built around two XPath expressions:
\itemize{
\item Multiline code blocks are identified by \verb{<pre>} tags with
class \code{sourceCode} and \code{r}, and are processed with \code{\link[=highlight]{highlight()}}.
\item Inline code is identified by \verb{<code>} tags that contain only text,
don't have a header tag (e.g. \verb{<h1>}) or \verb{<a>} as an ancestor,
and are porcessed processed with \code{\link[=autolink]{autolink()}}.
}

Use \code{downlit_html()} to process an \code{.html} file on disk;
use \code{downlit_xml_node()} to process an \code{xml_node} as part of a larger
pipeline.
}
