% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{href_topic}
\alias{href_topic}
\alias{href_article}
\title{Generate url for topic/article}
\usage{
href_topic(topic, package = NULL)

href_article(article, package = NULL)
}
\arguments{
\item{topic, article}{Topic/article name}

\item{package}{Optional package name}
}
\description{
Generate url for topic/article
}
\keyword{internal}
