% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme-document.R
\name{readme_document}
\alias{readme_document}
\title{Auto-linking for README files}
\usage{
readme_document(...)
}
\arguments{
\item{...}{Passed on to \link[rmarkdown:github_document]{rmarkdown::github_document}}
}
\description{
A drop-in replacement for \link[rmarkdown:github_document]{rmarkdown::github_document} that applies
highlighting and auto-linking to the resulting Markdown file.
GitHub strips the highlighting when displaying the README file,
it is still available when the site is displayed in pkgdown.
}
\examples{

# Include this into the front matter of your README.Rmd:
#
# output: downlit::readme_document
}
