% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{register_default_package_ref}
\alias{register_default_package_ref}
\alias{get_default_package_ref}
\title{Get or register the default package reference function}
\usage{
register_default_package_ref(fn)

get_default_package_ref()
}
\arguments{
\item{fn}{Function to be used to generate the package reference URL. The function should accept a single
parameter (preferably \code{package}) and return a character string of length one.}
}
\value{
This function is called for its side-effects, and so returns NULL invisibly.

Returns the currently registered function.
}
\description{
Get or register the default package reference function
}
\section{Functions}{
\itemize{
\item \code{register_default_package_ref}: Register the default package reference function
Register a function that will be used to generate the package reference URL if a remote URL for the package cannot be found.
By default, the \code{rdrr_package_ref()} function is used, using the "http://rdrr.io" website.

\item \code{get_default_package_ref}: Get the currently registered default package reference function
}}

\examples{
new_ref_function <- function(package) {
   paste0("https://rdocumentationsite/", package)
}
register_default_package_ref(new_ref_function)
autolink("stats::median()")
get_default_package_ref()
}
