% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{register_package_ref_function}
\alias{register_package_ref_function}
\alias{get_package_ref_function}
\alias{reset_package_ref_function}
\title{Get or register the default package reference function}
\usage{
register_package_ref_function(fn)

get_package_ref_function()

reset_package_ref_function()
}
\arguments{
\item{fn}{Function to be used to generate the package reference URL. The function should accept a single
parameter (preferably \code{package}) and return a character string of length one.}
}
\value{
\code{register_package_ref_function()} returns NULL invisibly.

\code{get_package_ref_function()} will return the currently registered function. By default, or if the registered function has been unset, this will return the
\code{rdrr_package_ref()} function which relies on the \url{https://rdrr.io} site.

\code{reset_package_ref_function()} returns NULL invisibly.
}
\description{
Register a function that will be used to generate the package reference URL if a remote URL for the package cannot be found.
By default, the \code{rdrr_package_ref()} function is used, relying the \url{https://rdrr.io} website.
}
\section{Functions}{
\itemize{
\item \code{register_package_ref_function}: Register the default package reference function

\item \code{get_package_ref_function}: Get the currently registered default package reference function

\item \code{reset_package_ref_function}: Resets the default package reference function to the \code{rdrr_package_ref()} function that
relies on the \url{https://rdrr.io} site.
}}

\examples{
new_ref_function <- function(package) {
   paste0("https://rdocumentationsite/", package)
}
register_package_ref_function(new_ref_function)
get_package_ref_function()
autolink("stats::median()")
}
