% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{replay_html}
\alias{replay_html}
\title{Convert object to HTML}
\usage{
replay_html(x, ...)
}
\arguments{
\item{x}{Object to display}

\item{...}{Additional parameters to pass to methods.}
}
\value{
character vector containing html
}
\description{
Convert object to HTML
}
