% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize_code}
\alias{tokenize_code}
\title{Perform tokenization of R code}
\usage{
tokenize_code(text, classes = classes_chroma())
}
\arguments{
\item{text}{String of code to highlight and link.}

\item{classes}{A mapping between token names and CSS class names.
Bundled \code{classes_pandoc()} and \code{classes_chroma()} provide mappings
that (roughly) match Pandoc and chroma (used by hugo) classes so you
can use existing themes.}
}
\value{
A list containing the input code as a string and as an expression
along with a data frame containing the tokenized code.
}
\description{
This function performs the initial step of the highlighting, turning R code
into a number of tokens identified by their type and attached a class based
on the \code{classes} argument.
}
\examples{
tokenize_code("sum(sample(10) * 10)")

}
\keyword{internal}
