% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{autolink}
\alias{autolink}
\title{Automatically link inline code}
\usage{
autolink(text)
}
\arguments{
\item{text}{String of code to highlight and link.}
}
\value{
If \code{text} is linkable, an HTML link. Otherwise, \code{NA}.
}
\description{
Automatically link inline code
}
\examples{
autolink("stats::median()")
autolink("vignette('grid', package = 'grid')")
}
