% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downlit-package.R
\docType{package}
\name{downlit-package}
\alias{downlit}
\alias{downlit-package}
\title{downlit: Syntax Highlighting and Automatic Linking}
\description{
Syntax highlight of R code, specifically designed
    for the needs of RMarkdown packages like pkgdown, hugodown, and
    bookdown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/downlit}
  \item Report bugs at \url{https://github.com/r-lib/downlit/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
