% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\alias{classes_pandoc}
\alias{classes_chroma}
\title{Highlight and link a code block}
\usage{
highlight(text, classes = classes_pandoc())

classes_pandoc()

classes_chroma()
}
\arguments{
\item{text}{String of code to highlight and link.}

\item{classes}{A mapping between token names and CSS class names.
Bundled \code{classes_pandoc()} and \code{classes_chroma()} provide mappings
that (roughly) match Pandoc and chroma (used by hugo) classes so you
can use existing themes.}
}
\value{
If \code{text} is valid R code, an HTML formatted string. Otherwise,
\code{NULL}.
}
\description{
This function:
\itemize{
\item syntax highlights code
\item links function calls to their documentation (where possible)
\item in comments, translates ANSI escapes in to HTML equivalents.
}
}
\examples{
cat(highlight("1 + 1"))
}
