% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\alias{classes_pandoc}
\alias{classes_chroma}
\title{Highlight and link a code block}
\usage{
highlight(text, classes = classes_chroma(), pre_class = NULL)

classes_pandoc()

classes_chroma()
}
\arguments{
\item{text}{String of code to highlight and link.}

\item{classes}{A mapping between token names and CSS class names.
Bundled \code{classes_pandoc()} and \code{classes_chroma()} provide mappings
that (roughly) match Pandoc and chroma (used by hugo) classes so you
can use existing themes.}

\item{pre_class}{Class(es) to give output \verb{<pre>}.}
}
\value{
If \code{text} is valid R code, an HTML \verb{<pre>} tag. Otherwise,
\code{NA}.
}
\description{
This function:
\itemize{
\item syntax highlights code
\item links function calls to their documentation (where possible)
\item in comments, translates ANSI escapes in to HTML equivalents.
}
}
\examples{
cat(highlight("1 + 1"))
cat(highlight("base::t(1:3)"))

# Unparseable R code returns NA
cat(highlight("base::t("))
}
